pyproject.toml
[project]
name = "apply-dataform-workflows"
version = "0.1.0"
description = "Apply Dataform release/workflow configurations from a JSON SSoT file"
requires-python = "==3.12.*"
dependencies = [
    "google-auth==2.49.1",
    "requests==2.32.5",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apply_dataform_workflows"]

[tool.pytest.ini_options]
testpaths = ["tests"]
